api/requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
configparser==3.5.0
cryptography==2.3.1
Django==2.1.3
django-filter==2.0.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
idna==2.7
pycparser==2.19
pytz==2018.7
six==1.11.0
mysqlclient==1.3.13
# Para intercambio de contenido en dominios cruzados
django-cors-headers
# Para solicitar los campos que deseemos de un serializer
drf-flex-fields
